+++ src/multisig/index.js
@@ -1,0 +1,27 @@
+// Copyright 2024 Tether Operations Limited
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+'use strict'
+
+/** @typedef {import('./wallet-account-read-only-multisig.js').MultisigInfo} MultisigInfo */
+/** @typedef {import('./wallet-account-read-only-multisig.js').MultisigProposal} MultisigProposal */
+/** @typedef {import('./wallet-account-read-only-multisig.js').MultisigMessage} MultisigMessage */
+
+/** @typedef {import('./wallet-account-multisig.js').MultisigTransactionResult} MultisigTransactionResult */
+/** @typedef {import('./wallet-account-multisig.js').MultisigTransactionOptions} MultisigTransactionOptions */
+/** @typedef {import('./wallet-account-multisig.js').MultisigMessageProposal} MultisigMessageProposal */
+/** @typedef {import('./wallet-account-multisig.js').MultisigOptions} MultisigOptions */
+
+export { IWalletAccountReadOnlyMultisig } from './wallet-account-read-only-multisig.js'
+
+export { IWalletAccountMultisig } from './wallet-account-multisig.js'
