+++ types/src/multisig/index.d.ts
@@ -1,0 +1,9 @@
+export { IWalletAccountReadOnlyMultisig } from "./wallet-account-read-only-multisig.js";
+export { IWalletAccountMultisig } from "./wallet-account-multisig.js";
+export type MultisigInfo = import("./wallet-account-read-only-multisig.js").MultisigInfo;
+export type MultisigProposal = import("./wallet-account-read-only-multisig.js").MultisigProposal;
+export type MultisigMessage = import("./wallet-account-read-only-multisig.js").MultisigMessage;
+export type MultisigTransactionResult = import("./wallet-account-multisig.js").MultisigTransactionResult;
+export type MultisigTransactionOptions = import("./wallet-account-multisig.js").MultisigTransactionOptions;
+export type MultisigMessageProposal = import("./wallet-account-multisig.js").MultisigMessageProposal;
+export type MultisigOptions = import("./wallet-account-multisig.js").MultisigOptions;
